move tests for pad_contrast_matrix to _utils
